se añade sistema notificacion toast

+++ tienda_videojuegos/carrito/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from catalogo.models import Juego
 from .models import Carrito, ItemCarrito
 
@@ -8,10 +9,14 @@ from .models import Carrito, ItemCarrito
 def agregar_al_carrito(request, juego_id):
     juego = get_object_or_404(Juego, id=juego_id)
     carrito, _ = Carrito.objects.get_or_create(usuario=request.user)
+    
     item, creado = ItemCarrito.objects.get_or_create(carrito=carrito, juego=juego)
     if not creado:
         item.cantidad += 1
-        item.save()
+        messages.info(request, f"Se ha aumentado la cantidad de '{juego.nombre}' en el carrito")
+    else:
+        messages.success(request, f"Producto '{juego.nombre}' agregado al carrito")
+    item.save()
 
     return redirect("catalogo:detalle_juego", pk=juego_id)
 
@@ -34,11 +39,18 @@ def eliminar_del_carrito(request, juego_id):
     item = carrito.items.filter(juego_id=juego_id).first()
     if item:
         item.delete()
+        messages.warning(request, f"Producto '{item.juego.nombre}' eliminado del carrito")
+    else:
+        messages.error(request, "El producto no se encontro en el carrito")
     return redirect("carrito:ver_carrito")
 
 
 @login_required
 def limpiar_carrito(request):
     carrito, _ = Carrito.objects.get_or_create(usuario=request.user)
-    carrito.items.all().delete()
+    if carrito.items.exists():
+        carrito.items.all().delete()
+        messages.warning(request, "Se han eliminado todos los productos del carrito")
+    else:
+        messages.info(request, "El carrito esta vacio")
     return redirect("carrito:ver_carrito")
